Compute fill_ones in unsigned space; remove dead SafeI64 code
- fill_ones: compute (UInt128(1) << bits) - 1 in unsigned to avoid
  signed overflow UB when bits == 127
- Remove dead SafeI64 overload of convert_NtoW_overflow (Weight = Number,
  so the Number& overload is always selected)
- Remove #include "arith/num_safeint.hpp" from zone_domain.hpp
- Update outdated Weight comment in splitdbm/definitions.hpp

+++ src/arith/num_big.hpp
@@ -436,7 +436,7 @@ class Number final {
         const auto u = static_cast<UInt128>(_n);
         const int bits = 128 - clz128(u);
         assert(bits > 0 && bits <= 64);
-        return Number{(static_cast<Int128>(1) << bits) - 1};
+        return Number{static_cast<Int128>((static_cast<UInt128>(1) << bits) - 1)};
     }
 
     friend std::ostream& operator<<(std::ostream& o, const Number& z);
